Added the ability to save an image capture of the current render
Added a function in render.c to save the current pixels from the
render target into an PNG file. Modified main.c to add a new key
binding for the letter P to save an image of the render. Added a
new linked flag for linking the SDL_image library.